requirements.txt
grpcio==1.62.1
protobuf==4.25.3
psycopg2-binary==2.9.9
